requirements.txt
pycryptodome
pyqrcode
fpdf
